pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.6.3",
    "langsmith>=0.4.10",
    "notebook>=7.4.4",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.65",
]
